Task/FP-1168: add backend unit tests for add/delete members of TAS projects (subtask of FP-737)  (#467)
* Remove unneeded change

* Fix url for user delete

* Remove logging statements and correct ApiException

Second argument of ApiException is status (i.e. status_code)

* Remove status code in ApiException as TAS returns 200

Main reason for not using TAS return code is that for errors (like
missing user), TAS still returns a 200.  Also, it isn't useful as
we handle all ApiException as 400 and don't log that status_code that
that was pass to ApiException

* Add tests

* Fix pep errors

* Remove change to response for testing purposes

* Fix use of ApiException

* Fix tests

* Improve logging message for forced allocation retrieval

* Fix logging for add/delete of users to allocation